+++ marshmallow/utils.py
@@ -2,6 +2,10 @@
 """Utility methods for marshmallow."""
 from __future__ import absolute_import
 import json
+import datetime
+from email.utils import formatdate
+from calendar import timegm
+
 from decimal import Decimal, Context, Inexact
 from pprint import pprint as py_pprint
 
@@ -70,3 +74,81 @@ def pprint(obj, *args, **kwargs):
         print(json.dumps(obj, *args, **kwargs))
     else:
         py_pprint(obj, *args, **kwargs)
+
+# From pytz: http://pytz.sourceforge.net/
+ZERO = datetime.timedelta(0)
+HOUR = datetime.timedelta(hours=1)
+
+
+class UTC(datetime.tzinfo):
+    """UTC
+
+    Optimized UTC implementation. It unpickles using the single module global
+    instance defined beneath this class declaration.
+    """
+    zone = "UTC"
+
+    _utcoffset = ZERO
+    _dst = ZERO
+    _tzname = zone
+
+    def fromutc(self, dt):
+        if dt.tzinfo is None:
+            return self.localize(dt)
+        return super(utc.__class__, self).fromutc(dt)
+
+    def utcoffset(self, dt):
+        return ZERO
+
+    def tzname(self, dt):
+        return "UTC"
+
+    def dst(self, dt):
+        return ZERO
+
+    def __reduce__(self):
+        return _UTC, ()
+
+    def localize(self, dt, is_dst=False):
+        '''Convert naive time to local time'''
+        if dt.tzinfo is not None:
+            raise ValueError('Not naive datetime (tzinfo is already set)')
+        return dt.replace(tzinfo=self)
+
+    def normalize(self, dt, is_dst=False):
+        '''Correct the timezone information on the given datetime'''
+        if dt.tzinfo is self:
+            return dt
+        if dt.tzinfo is None:
+            raise ValueError('Naive time - no tzinfo set')
+        return dt.astimezone(self)
+
+    def __repr__(self):
+        return "<UTC>"
+
+    def __str__(self):
+        return "UTC"
+
+UTC = utc = UTC()  # UTC is a singleton
+
+
+def rfcformat(dt, localtime=False):
+    '''Return the RFC822-formatted represenation of a datetime object.
+
+    :param bool localtime: If ``True``, return the date relative to the local
+        timezone instead of UTC, properly taking daylight savings time into account.
+    '''
+    return formatdate(timegm(dt.utctimetuple()), localtime=localtime)
+
+
+def isoformat(dt, localtime=False, *args, **kwargs):
+    '''Return the ISO8601-formatted UTC representation of a datetime object.
+    '''
+    if localtime and dt.tzinfo is not None:
+        localized = dt
+    else:
+        if dt.tzinfo is None:
+            localized = UTC.localize(dt)
+        else:
+            localized = dt.astimezone(UTC)
+    return localized.isoformat(*args, **kwargs)